Update ExtractTelos.py to Python 3.10
Updated the Python version and made the necessary alterations. The bigger one as with the use of the set() function, I needed to change the code for some loops, because the set() function was not working with SeqRecord object type. I did some tests and the results were the same as the previous code.